feat(events-v2): Add better styling to events table
Also moves the hardcoded data that drives the 3 events views into a
separate file.
