+++ tests/js/spec/views/organizationEventsV2/utils.spec.jsx
@@ -1,4 +1,5 @@
-import {ALL_VIEWS, getCurrentView, getQuery} from 'app/views/organizationEventsV2/utils';
+import {getCurrentView, getQuery} from 'app/views/organizationEventsV2/utils';
+import {ALL_VIEWS} from 'app/views/organizationEventsV2/data';
 
 describe('getCurrentView()', function() {
   it('returns current view', function() {
